requirements.txt
agency-swarm[fastapi]>=1.2.1
fastapi
uvicorn
python-docx
PyMuPDF
python-pptx
fpdf2
Pillow
requests
